editar y eliminar completo

+++ controllers/event.controller.js
@@ -82,7 +82,10 @@ module.exports.deleteById = (req, res) => {
     const imageUrl = event.imageUrl;
     const imageName = imageUrl.split("/").pop();
     const file = bucket.file(imageName);
-    file.delete();
+    if(file){
+      file.delete();
+
+    }
   });
 
   Event.findByIdAndDelete(id)
@@ -90,4 +93,103 @@ module.exports.deleteById = (req, res) => {
     .catch((err) =>
       res.status(400).json({ msg: "Error al intentar eliminar el evento", err })
     );
+};
+
+module.exports.updateOneEvent = (req, res) => {
+  const { name, type, description, startDate, endDate, price, location } = req.body;
+  const eventImage = req.file;
+  const {id} = req.params
+
+  Event.findById(id)
+    .then(eventoExistente => {
+      if (!eventoExistente) {
+        return res.status(404).send("Evento no encontrado");
+      }
+
+      let imageUrl = eventoExistente.imageUrl;
+
+      if (eventImage) {
+        
+        const oldImageName = imageUrl.split('/').pop();
+        const oldImageFile = bucket.file(oldImageName);
+
+        oldImageFile.delete()
+          .then(() => {
+            
+            const blob = bucket.file(eventImage.originalname);
+            const blobStream = blob.createWriteStream({
+              metadata: {
+                contentType: eventImage.mimetype,
+              },
+            });
+
+            blobStream.on("error", (err) => {
+              console.error(err);
+              res.status(500).send("Error al subir la nueva imagen");
+            });
+
+            blobStream.on("finish", () => {
+              blob.makePublic()
+                .then(() => {
+                  imageUrl = `https://storage.googleapis.com/${bucket.name}/${blob.name}`;
+
+                  
+                  Object.assign(eventoExistente, {
+                    name,
+                    type,
+                    description,
+                    startDate,
+                    endDate,
+                    price,
+                    location,
+                    imageUrl
+                  });
+
+                  eventoExistente.save()
+                    .then(() => {
+                      res.status(200).send("Evento actualizado exitosamente con nueva imagen");
+                    })
+                    .catch(error => {
+                      console.error(error);
+                      res.status(500).send("Error al actualizar el evento");
+                    });
+                })
+                .catch(error => {
+                  console.error(error);
+                  res.status(500).send("Error al hacer pública la imagen");
+                });
+            });
+
+            blobStream.end(eventImage.buffer);
+          })
+          .catch(error => {
+            console.error(error);
+            res.status(500).send("Error al eliminar la imagen antigua");
+          });
+      } else {
+        
+        Object.assign(eventoExistente, {
+          name,
+          type,
+          description,
+          startDate,
+          endDate,
+          price,
+          location
+        });
+
+        eventoExistente.save()
+          .then(() => {
+            res.status(200).send("Evento actualizado exitosamente sin cambiar imagen");
+          })
+          .catch(error => {
+            console.error(error);
+            res.status(500).send("Error al actualizar el evento");
+          });
+      }
+    })
+    .catch(error => {
+      console.error(error);
+      res.status(500).send("Error en el servidor");
+    });
 };
